// WADNR.Web/src/app/shared/generated/model/project-basics-step-request.ts
export class ProjectBasicsStepRequest { 
    ProjectName?: string | null;
    ProjectDescription?: string | null;
    ProjectTypeID?: number;
    ProjectStageID?: number;
    PlannedDate?: string | null;
    CompletionDate?: string | null;
    ExpirationDate?: string | null;
    FocusAreaID?: number | null;
    LeadImplementerOrganizationID?: number | null;
    PercentageMatch?: number | null;
    ProgramIDs?: Array<number> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}