// WADNR.Web/src/app/shared/generated/model/program-grid-row.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrganizationLookupItem } from './organization-lookup-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProgramGridRow { 
    ProgramID?: number;
    ProgramName?: string | null;
    ProgramShortName?: string | null;
    IsActive?: boolean;
    IsDefaultProgramForImportOnly?: boolean;
    Organization?: OrganizationLookupItem;
    ProjectCount?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProgramGridRowForm { 
    ProgramID?: FormControl<number>;
    ProgramName?: FormControl<string>;
    ProgramShortName?: FormControl<string>;
    IsActive?: FormControl<boolean>;
    IsDefaultProgramForImportOnly?: FormControl<boolean>;
    Organization?: FormControl<OrganizationLookupItem>;
    ProjectCount?: FormControl<number>;
}

export class ProgramGridRowFormControls { 
    public static ProgramID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramShortName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsActive = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsDefaultProgramForImportOnly = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Organization = (value: FormControlState<OrganizationLookupItem> | OrganizationLookupItem = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<OrganizationLookupItem>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectCount = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}